+++ 1-practica-cloud-inventario-rest/src/main/java/com/everis/practicacloudinventario/service/InventarioService.java
@@ -12,7 +12,7 @@ public class InventarioService {
 	@Autowired
 	private InventarioRepository inventarioRepository;
 	
-	public Inventario consultarInventario(Long id) {
+	public Inventario consultarInventario(int id) {
 		return inventarioRepository.findById(id).get();
 	}
